+++ src/types/rx-document.d.ts
@@ -19,11 +19,11 @@ export type RxDocument<RxDocumentType = {}, OrmMethods = {}> = RxDocumentBase<Rx
 
 
 /**
- * The public facing atomic update function.
+ * The public facing modify update function.
  * It only gets the document parts as input, that
  * are mutateable by the user.
  */
-export type AtomicUpdateFunction<RxDocumentType> = (
+export type ModifyFunction<RxDocumentType> = (
     doc: WithDeleted<RxDocumentType>
 ) => MaybePromise<WithDeleted<RxDocumentType>> | MaybePromise<RxDocumentType>;
 
@@ -59,7 +59,6 @@ export declare interface RxDocumentBase<RxDocType, OrmMethods = {}> {
     _data: RxDocumentData<RxDocType>;
     primaryPath: string;
     revision: string;
-    _atomicQueue: Promise<any>;
     $emit(cE: RxChangeEvent<RxDocType>): void;
     _saveData(newData: any, oldData: any): Promise<RxDocument<RxDocType, OrmMethods>>;
     // /internal things
@@ -75,15 +74,22 @@ export declare interface RxDocumentBase<RxDocType, OrmMethods = {}> {
     /**
      * mutate the document with a function
      */
-    atomicUpdate(mutationFunction: AtomicUpdateFunction<RxDocType>, context?: string): Promise<RxDocument<RxDocType, OrmMethods>>;
+    modify(mutationFunction: ModifyFunction<RxDocType>, context?: string): Promise<RxDocument<RxDocType, OrmMethods>>;
+    incrementalModify(mutationFunction: ModifyFunction<RxDocType>, context?: string): Promise<RxDocument<RxDocType, OrmMethods>>;
+
     /**
      * patches the given properties
      */
-    atomicPatch(patch: Partial<RxDocType>): Promise<RxDocument<RxDocType, OrmMethods>>;
+    patch(patch: Partial<RxDocType>): Promise<RxDocument<RxDocType, OrmMethods>>;
+    incrementalPatch(patch: Partial<RxDocType>): Promise<RxDocument<RxDocType, OrmMethods>>;
 
     update(updateObj: UpdateQuery<RxDocType>): Promise<RxDocument<RxDocType, OrmMethods>>;
+    incrementalUpdate(updateObj: UpdateQuery<RxDocType>): Promise<RxDocument<RxDocType, OrmMethods>>;
+
     updateCRDT(updateObj: CRDTEntry<RxDocType> | CRDTEntry<RxDocType>[]): Promise<RxDocument<RxDocType, OrmMethods>>;
+
     remove(): Promise<RxDocument<RxDocType, OrmMethods>>;
+    incrementalRemove(): Promise<RxDocument<RxDocType, OrmMethods>>;
 
     // only for temporary documents
     set(objPath: string, value: any): RxDocument<RxDocType, OrmMethods>;