+++ test/unit/reactive-collection.test.ts
@@ -170,8 +170,8 @@ config.parallel('reactive-collection.test.js', () => {
             await c.insert(schemaObjects.human());
             await doc3.remove();
 
-            await doc1.atomicPatch({ firstName: 'foobar1' });
-            await doc2.atomicPatch({ firstName: 'foobar2' });
+            await doc1.incrementalPatch({ firstName: 'foobar1' });
+            await doc2.incrementalPatch({ firstName: 'foobar2' });
 
             await AsyncTestUtil.waitUntil(() => emitted.length === 2);
             emitted.forEach(cE => assert.strictEqual(cE.operation, 'UPDATE'));