+++ test/unit/replication-p2p.test.ts
@@ -169,7 +169,7 @@ describe('replication-p2p.test.ts', () => {
 
             // update
             const doc = await c1.findOne().exec(true);
-            await doc.getLatest().atomicPatch({ age: 100 });
+            await doc.getLatest().incrementalPatch({ age: 100 });
             await awaitCollectionsInSync([c1, c2]);
             assert.strictEqual(doc.getLatest().age, 100);
             await wait(100);